+++ src/grammar.pest
@@ -5,7 +5,7 @@ ACTIVITY_DIAGRAM = {start_activity ~ new_line+ ~ ACTIVITY_BODY}
 ACTIVITY_BODY = {start_state ~ (PATH ~ new_line)+ ~ new_line?}
 PATH = {(ACTIVITY | IF)+ ~ END_STATE?}
 END_STATE = {solid_arrow ~ end_state}
-ACTIVITY = {solid_arrow ~ (label | identifier) ~ (colon ~ (label | identifier))? ~ new_line?}
+ACTIVITY = {solid_arrow ~ (label | identifier) ~ (colon ~ (label | identifier))? ~ new_line+?}
 IF = {IF_STATEMENT ~ WHITE_SPACE? ~ ELSE_STATEMENT ~ new_line?}
 IF_STATEMENT = {if_kw ~ WHITESPACE? ~ identifier ~ WHITESPACE? ~ left_bracket ~ new_line ~ PATH ~ new_line? ~ right_bracket ~ WHITESPACE?}
 ELSE_STATEMENT = {else_kw ~ WHITESPACE? ~ left_bracket ~ new_line ~ PATH ~ new_line? ~ right_bracket}
