+++ src/rules/mod.rs
@@ -4,3 +4,6 @@ pub mod context;
 pub mod use_case;
 pub mod structs;
 pub mod activity;
+pub mod node;
+pub mod activity_utils;
+pub mod path;
